Move BlockInspectorButton component to separate file
